requirements.txt
# =============================================================================
# HUB FINANCEIRO - DEPENDÊNCIAS DO PROJETO
# =============================================================================

# =============================================================================
# DJANGO E FRAMEWORK BASE
# =============================================================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-import-export==3.3.1
django-extensions==3.2.3

# =============================================================================
# AUTENTICAÇÃO E SEGURANÇA
# =============================================================================
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7
PyJWT==2.8.0

# =============================================================================
# BANCO DE DADOS
# =============================================================================
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1

# =============================================================================
# PROCESSAMENTO ASSÍNCRONO
# =============================================================================
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
kombu==5.3.4

# =============================================================================
# WEBSOCKETS E COMUNICAÇÃO EM TEMPO REAL
# =============================================================================
channels==4.0.0
channels-redis==4.1.0
websockets==12.0

# =============================================================================
# APIS E INTEGRAÇÕES
# =============================================================================
requests==2.31.0
urllib3==2.1.0
httpx==0.25.2

# =============================================================================
# TELEGRAM BOT
# =============================================================================
python-telegram-bot==20.7
aiogram==3.3.0

# =============================================================================
# WHATSAPP INTEGRATION
# =============================================================================
twilio==8.10.3
whatsapp-web.py==0.3.3

# =============================================================================
# DADOS FINANCEIROS E ANÁLISE
# =============================================================================
yfinance==0.2.28
alpha-vantage==2.3.1
polygon-api-client==1.13.2
quandl==3.7.0
financetoolkit==1.8.5
investpy==1.0.8

# =============================================================================
# ANÁLISE DE DADOS E MACHINE LEARNING
# =============================================================================
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
ta-lib==0.4.28
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# =============================================================================
# ANÁLISE TÉCNICA
# =============================================================================
TA-Lib==0.4.28
pandas-ta==0.3.14b0
backtrader==1.9.78.123
zipline==3.0.0

# =============================================================================
# INTELIGÊNCIA ARTIFICIAL
# =============================================================================
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
langchain==0.0.350
llama-index==0.9.15

# =============================================================================
# PROCESSAMENTO DE TEXTO E NLP
# =============================================================================
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2
beautifulsoup4==4.12.2

# =============================================================================
# UTILIDADES
# =============================================================================
python-decouple==3.8
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9

# =============================================================================
# VALIDAÇÃO E FORMATAÇÃO
# =============================================================================
marshmallow==3.20.1
pydantic==2.5.1
validators==0.22.0

# =============================================================================
# TESTES
# =============================================================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# =============================================================================
# DEPLOY E PRODUÇÃO
# =============================================================================
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# =============================================================================
# MONITORAMENTO E LOGGING
# =============================================================================
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0
django-silk==5.0.4

# =============================================================================
# CACHE E PERFORMANCE
# =============================================================================
django-cachalot==2.6.1
django-compressor==4.4

# =============================================================================
# NOTIFICAÇÕES
# =============================================================================
pyfcm==1.5.4
plyer==2.1.0

# =============================================================================
# MANIPULAÇÃO DE ARQUIVOS
# =============================================================================
pypdf2==3.0.1
python-docx==1.1.0
csvkit==1.1.1

# =============================================================================
# TIMEZONE E DATAS
# =============================================================================
pytz==2023.3
python-dateutil==2.8.2

# =============================================================================
# NETWORKING E COMUNICAÇÃO
# =============================================================================
aiohttp==3.9.1
websocket-client==1.6.4

# =============================================================================
# DESENVOLVIMENTO
# =============================================================================
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0

# =============================================================================
# DOCUMENTAÇÃO
# =============================================================================
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# =============================================================================
# EXPORT E RELATÓRIOS
# =============================================================================
weasyprint==60.2
jinja2==3.1.2